requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
httpx>=0.25.0
tiktoken>=0.5.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.0
transformers>=4.40.0
torch>=2.0.0
pillow>=10.0.0
accelerate>=0.20.0

